+++ internal/proposal/featured.go
@@ -1,0 +1,18 @@
+package proposal
+
+import (
+	"time"
+)
+
+type Featured struct {
+	ProposalID string
+
+	CreatedAt time.Time
+
+	StartAt time.Time
+	EndAt   time.Time
+}
+
+func (Featured) TableName() string {
+	return "featured_proposals"
+}